+++ packages/json-rpc/ws-middleware/src/hybrid.ts
@@ -1,0 +1,103 @@
+import { JsonRpcMessage } from "@polkadot-api/json-rpc-provider"
+import { filter, Subject, Subscription, take } from "rxjs"
+import { withLegacy } from "./legacy"
+import { modern } from "./modern"
+import { Middleware } from "./types"
+
+const modernGroups = ["chainHead", "transaction", "chainSpec", "archive"].map(
+  (name) => `${name}_v1`,
+)
+const isModern = (methods: string[]): boolean =>
+  modernGroups.every((group) => methods.some((m) => m.startsWith(group)))
+
+export const hybridMiddleware = (methods: string[]): Middleware => {
+  if (isModern(methods)) return modern
+
+  const individualChecks = Object.fromEntries(
+    modernGroups.map((group) => [
+      group,
+      !!methods.find((v) => v.startsWith(group)),
+    ]),
+  )
+
+  return (base) => {
+    const multiplexed = multiplex(base)
+    const modernProvider = modern(multiplexed)
+    const legacyProvider = withLegacy(multiplexed)
+
+    return (onMsg, onHalt) => {
+      const modernConnection = modernProvider((message) => {
+        if ("error" in message && message.error.code === -32800) {
+          console.warn("Max follow connections received: chainHead_v1 disabled")
+          individualChecks.chainHead_v1 = false
+        }
+        onMsg(message)
+      }, onHalt)
+      const legacyConnection = legacyProvider(onMsg, onHalt)
+
+      return {
+        send(message) {
+          for (const group of modernGroups) {
+            if (message.method.startsWith(group)) {
+              if (individualChecks[group]) {
+                modernConnection.send(message)
+              } else {
+                legacyConnection.send(message)
+              }
+              return
+            }
+          }
+          modernConnection.send(message)
+        },
+        disconnect() {
+          modernConnection.disconnect()
+          legacyConnection.disconnect()
+        },
+      }
+    }
+  }
+}
+
+const multiplex: Middleware = (base) => {
+  const halt$ = new Subject<any>()
+  const msg$ = new Subject<JsonRpcMessage>()
+
+  const baseConnection = base(
+    (msg) => msg$.next(msg),
+    (e) => halt$.next(e),
+  )
+
+  return (onMsg, onHalt) => {
+    const notificationSub = msg$
+      .pipe(filter((v) => v.id == null))
+      .subscribe(onMsg)
+    const haltSub = halt$.subscribe(onHalt)
+    const responseSubs = new Set<Subscription>()
+
+    return {
+      send(message) {
+        if (message.id == null) {
+          return baseConnection.send(message)
+        }
+        const id = message.id
+        const responseSub = msg$
+          .pipe(
+            filter((r) => r.id === id),
+            take(1),
+          )
+          .subscribe((msg) => {
+            onMsg(msg)
+            if (!responseSub) throw new Error("unreachable")
+            responseSubs.delete(responseSub)
+          })
+        responseSubs.add(responseSub)
+        baseConnection.send(message)
+      },
+      disconnect() {
+        notificationSub.unsubscribe()
+        haltSub.unsubscribe()
+        responseSubs.forEach((s) => s.unsubscribe())
+      },
+    }
+  }
+}