+++ packages/json-rpc/ws-middleware/src/middleware.ts
@@ -1,18 +1,6 @@
-import { withLegacy } from "./legacy"
+import { hybridMiddleware } from "./hybrid"
 import { methodsRouter } from "./methods-router"
-import { modern } from "./modern"
 import { withNumericIds } from "./numeric-ids"
 import { apply } from "./utils"
 
-const modernGroups = [
-  "chainHead",
-  "transaction",
-  "chainSpec" /*, "archive"*/,
-].map((name) => `${name}_v1`)
-const isModern = (methods: string[]): boolean =>
-  modernGroups.every((group) => methods.some((m) => m.startsWith(group)))
-
-export const middleware = apply(
-  withNumericIds,
-  methodsRouter((methods) => (isModern(methods) ? modern : withLegacy)),
-)
+export const middleware = apply(withNumericIds, methodsRouter(hybridMiddleware))