+++ packages/client/CHANGELOG.md
@@ -7,6 +7,9 @@
 - **codegen**
   - Use all metadata entries for whitelist type utilities
 
+- **WS-Provider:**
+  - Use granular legacy/modern enhancers based on groups of functions
+
 ## 2.0.0-rc.5 - 2026-03-13
 
 ### Fixed